// generated/src/app/api-models/claim-audit-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimAuditSelect {
  lineOfBusiness: string;
  claimType: string;
  claimAuditOrder: number;
  searchSequence: number;
  columnName: string;
  determinantType: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  claimAuditToolCode: number;
}